Building failed on X86 model

+++ source/hunt/collection/ByteBuffer.d
@@ -1026,7 +1026,7 @@ abstract class ByteBuffer : Buffer {
         return i + offset;
     }
 
-    protected long byteOffset(long i) {
+    protected size_t byteOffset(size_t i) {
         return offset + i;
     }
